cleanup: JWT 모두 삭제
