--- src/main/java/net/detalk/api/auth/domain/AuthToken.java
@@ -1,7 +1,0 @@
-package net.detalk.api.auth.domain;
-
-public record AuthToken (
-    String accessToken,
-    String refreshToken
-) {
-}
