--- src/main/java/net/detalk/api/auth/repository/AuthRefreshTokenRepository.java
@@ -1,11 +1,0 @@
-package net.detalk.api.auth.repository;
-
-import net.detalk.api.auth.domain.AuthRefreshToken;
-
-import java.util.Optional;
-
-public interface AuthRefreshTokenRepository {
-    AuthRefreshToken save(AuthRefreshToken token);
-    Optional<AuthRefreshToken> findByToken(String token);
-    AuthRefreshToken update(AuthRefreshToken token);
-}
